Déplacere classes communes dans paris.py
