+++ openfisca_france/model/prestations/prestations_locales/paris/paris_logement.py
@@ -41,11 +41,6 @@ class personnes_handicap_paris(Variable):
 		condition_aah = aah>0
 		return period,condition_aah
 
-class parisien(Variable):
-        column = BoolCol
-        entity_class = Familles
-        label = u"Résidant à Paris au moins 3 ans dans les 5 dernières années"
-
 class paris_logement_elig(Variable):
 	column=BoolCol
 	label=u"Personne qui est eligible pour l'aide PL"
@@ -90,66 +85,6 @@ class condition_taux_effort(Variable):
 		condition_loyer=loyer>=(ressources_mensuelles*0.3)
 		return period,condition_loyer
 
-
-class paris_logement_familles_br_i(Variable):
-    column = FloatCol
-    label = u"Base de ressources individuelle pour Paris Logement Famille"
-    entity_class = Individus
-
-    def function(self, simulation, period):
-        period = period.this_month
-        last_year = period.last_year
-        salaire_net = simulation.calculate('salaire_net', period)
-        chonet = simulation.calculate('chonet', period)
-        rstnet = simulation.calculate('rstnet', period)
-        pensions_alimentaires_percues = simulation.calculate('pensions_alimentaires_percues', period)
-        pensions_alimentaires_versees_individu = simulation.calculate(
-            'pensions_alimentaires_versees_individu', period)
-        rsa_base_ressources_patrimoine_i = simulation.calculate_add('rsa_base_ressources_patrimoine_i', period)
-        indemnites_journalieres_imposables = simulation.calculate('indemnites_journalieres_imposables', period)
-        indemnites_stage = simulation.calculate('indemnites_stage', period)
-        revenus_stage_formation_pro = simulation.calculate('revenus_stage_formation_pro', period)
-        allocation_securisation_professionnelle = simulation.calculate(
-            'allocation_securisation_professionnelle', period)
-        prestation_compensatoire = simulation.calculate('prestation_compensatoire', period)
-        pensions_invalidite = simulation.calculate('pensions_invalidite', period)
-        indemnites_chomage_partiel = simulation.calculate('indemnites_chomage_partiel', period)
-        bourse_recherche = simulation.calculate('bourse_recherche', period)
-        gains_exceptionnels = simulation.calculate('gains_exceptionnels', period)
-
-        def revenus_tns():
-            revenus_auto_entrepreneur = simulation.calculate_add('tns_auto_entrepreneur_benefice', period)
-
-            # Les revenus TNS hors AE sont estimés en se basant sur le revenu N-1
-            tns_micro_entreprise_benefice = simulation.calculate('tns_micro_entreprise_benefice', last_year) / 12
-            tns_benefice_exploitant_agricole = simulation.calculate('tns_benefice_exploitant_agricole', last_year) / 12
-            tns_autres_revenus = simulation.calculate('tns_autres_revenus', last_year) / 12
-
-            return revenus_auto_entrepreneur + tns_micro_entreprise_benefice + tns_benefice_exploitant_agricole + tns_autres_revenus
-
-        result = (
-            salaire_net + indemnites_chomage_partiel + indemnites_stage + chonet + rstnet +
-            pensions_alimentaires_percues - abs_(pensions_alimentaires_versees_individu) +
-            rsa_base_ressources_patrimoine_i + allocation_securisation_professionnelle +
-            indemnites_journalieres_imposables + prestation_compensatoire +
-            pensions_invalidite + bourse_recherche + gains_exceptionnels + revenus_tns() +
-            revenus_stage_formation_pro
-            )
-
-        return period, result
-
-class paris_logement_familles_br(Variable):
-    column = FloatCol
-    label = u"Base de ressource pour Paris Logement Famille"
-    entity_class = Familles
-
-    def function(self, simulation, period):
-        period = period.this_month
-        paris_logement_familles_br_i_holder = simulation.compute('paris_logement_familles_br_i', period)
-        paris_logement_familles_br = self.sum_by_entity(paris_logement_familles_br_i_holder)
-        result = paris_logement_familles_br
-
-        return period, result
 
 # Critères relatifs à PSOL (PA et PH)
